Moved authentication inside each microservice

+++ backend/ProductMicroService/src/index.ts
@@ -51,7 +51,6 @@ const protectedRout = (req: Request, res: Response) => {
         return ERROR_401;
     }
     let cookies = req.headers.cookie.split('; ');
-    console.log(cookies);
 
     // We get the token value from cookies.
     if (cookies.filter(str => str.startsWith("token")).length != 1) {
@@ -67,7 +66,7 @@ const protectedRout = (req: Request, res: Response) => {
 
     // Verify JWT token
     const user = verifyJWT(token);
-    console.log(`my name is ${user}`)
+    console.log(`my name is ${user.userId}`)
     if (!user) {
         res.statusCode = 401;
         res.end(
@@ -100,14 +99,13 @@ app.use(async (req: RequestWithPermission, res, next) => {
 
     const user = protectedRout(req, res);
     if (user != ERROR_401) {
-
+        console.log("getting permission..");
         await axios
             .get(`${userServiceURL}/api/user/${user.userId}/permission`, {
-                data: req.body,
                 headers: req.headers
             }).then(response => {
-                console.log(response)
-                req.permission = response.data;
+                console.log("received permission..");
+                req.permission = response.data.permission;
                 next();
             })
             .catch((error) => {
@@ -168,11 +166,14 @@ const getAllProducts = async (req: Request, res: Response, idOrType: string) => 
 
 
 const createProduct = async (req: RequestWithPermission, res: Response) => {
+    console.log('Creating a product');
+    console.log(req.permission);
     if (["A", "M"].includes(req.permission)) {
         // Parse request body as JSON
         try {
-            let { name, category, description, price, stock, image } = JSON.parse(req.body);
-            if (!name || !category || !description || !price || !stock || typeof price != 'number' || typeof stock != 'number' || !Number.isInteger(stock) || stock < 0 || price < 0 || price > 1000 || !validCategories.includes(category)) {
+            let { name, category, description, price, stock, image } = req.body;
+            console.log(req.body);
+            if (!name || !category || !description || !price || !stock || !image || typeof price != 'number' || typeof stock != 'number' || !Number.isInteger(stock) || stock < 0 || price < 0 || price > 1000 || !validCategories.includes(category)) {
 
                 res.writeHead(400, { 'Content-Type': 'application/json' });
                 res.end(JSON.stringify({ message: 'Invalid Details' }));
@@ -215,6 +216,7 @@ const createProduct = async (req: RequestWithPermission, res: Response) => {
 
 //TODO: #8 replace parsing body manually with express body-parser
 const updateProduct = async (req: RequestWithPermission, res: Response, id: string) => {
+    console.log("Updating Product...");
     if (["A", "M"].includes(req.permission)) {
         let prod;
         try {
@@ -232,9 +234,12 @@ const updateProduct = async (req: RequestWithPermission, res: Response, id: stri
         else {
 
             try {
+                console.log(req.body);
                 const { name, category, description, price, stock, image } = req.body;
                 if ((!name && !category && !description && !price && !stock && !image) ||
-                    typeof price != 'number' || typeof stock != 'number' || !Number.isInteger(stock) || stock < 0 || price < 0 || price > 1000 || !validCategories.includes(category)) {
+                    (price && (typeof price != 'number' || price < 0 || price > 1000)) ||
+                    (stock && (typeof stock != 'number' || !Number.isInteger(stock)) || stock < 0)
+                    || (category && !validCategories.includes(category))) {
 
                     res.writeHead(400, { 'Content-Type': 'application/json' });
                     res.end(JSON.stringify({ message: 'Invalid Details' }));
@@ -269,6 +274,12 @@ const removeProduct = async (req: RequestWithPermission, res: Response, id: stri
 
     if (req.permission == "A") {
         try {
+            let prod = await productService.getProductById(id);
+            if (!prod) {
+                res.statusCode = 404;
+                res.end();
+                return;
+            }
             await productService.removeProduct(id);
         } catch (err) {
             res.statusCode = 400;
